Operation Note Ops Numbering

+++ libraries/chain/include/steem/chain/database.hpp
@@ -522,7 +522,7 @@ namespace steem { namespace chain {
 
          transaction_id_type           _current_trx_id;
          uint32_t                      _current_block_num    = 0;
-         uint16_t                      _current_trx_in_block = 0;
+         int32_t                       _current_trx_in_block = 0;
          uint16_t                      _current_op_in_trx    = 0;
          uint16_t                      _current_virtual_op   = 0;
